requirements.txt
pytest

flask
flask_restx

pyyaml

spotipy

azapi

symspellpy

nltk
pronouncing

youtube-search-python
youtube_dl

pydub
moviepy